Created Java package/folder, Email class/cookie cutter, EmailApp which will have the objects
I am still using the HELLOJAVA folder and inside this folder I add two java files, one is the Email.java file and the other is the EmailApp.java file. The Email class is the cookie cutter which will use a constructor which is the process of cutting the cookie by placing the cutter on the dough and construct objects which is the purpose of the EmailApp, creating objects, different kinds of Email objects or apps. Note: A java package is nothing more than a folder.