Encapsulation, Constructor
Encapsulation in java is a process of wrapping code and data together into a single unit. We can create a fully encapsulated class in Java by making all the data members of the class private. The firstname, lastname, password, department, mailboxcapacity and alternateEmail are all set to private. Sometimes a setter and getter method is used. Advantages of encapsulation is that it provides only a setter or getter method, you can make a the class read-only or write-only. It provides control over the data. I can write the logic inside the setter method. It is a way to achieve data hiding in Java because other class will not be able to access the data through the private data members. It is better for unit testing. There are two types of modifiers in java: access modifiers and non-access modifiers. There are 4 types of java access modifiers: private, default, protected, public. The non-access modifiers are static, abstract, synchronized, native, volatile, transcient. The access modifiers in java specifies accessibility (scope) of a data member, method, constructor or class. The Email constructor with parameters firstName, and lastName. A constructor is a block of codes similar to the method. It is called when an instance of the object is create, and memory is allocated for the object. It is a special type of method which is uded to initialize the object