em1 is the new email object created in the EmailApp.java file.
em1 is the new object created in the EmailApp.java file. Using the System.out.printlin I output the current active firstName and lastName. When em1 object is created Polymorphism is used. Using IS-A can be used, em1 IS-A Email. Any Java object (em1) that can pass more than one IS-A test is considered to be polymorphic. em1 and em2 IS-A Email.