conditional statements
An 'if' statement comes with a boolean express in the parentheses and the statements which can be more than one in the brackets. If the boolean expression evaluates to true then the block of code or statement(s) in the bracket is executed. If the expression is not true then the statement after the closing bracket is executed. Now in the private string type setDepartment method there is an output or print to the system asking for input from the client which includes adding Scanner but before  importing java.util.Scanner in order for it to work. The if statement has a 'depChoice =1' expression and if it is true the expression returns 'sales' if not it goes to the 'else if' if that expression is true a 'dev' commit is returned. The nested if statement has one if, two else if, and 1 else that returns space, an empty space.