setDepartment method
In the constructor setDapartment is stored in this.department. The keyword 'this' can be used to refer current class instance variable, can be used to invoke current class method, this can be used to invoke current class constructor. 'this' can be used to refer current class instance variable. If there is ambuiguity between the instance variables and parameters, this keyword resolves the problem of ambiguity. The System.out.println prints out 'Department' and the current value in department. The setDepartment method has the if, else if, and else statement and Scanner and the print out