The Student object file
First start by creating a 'StudentManagaementSystem' folder or package in Java. In that package I create two java files. The Student.java file is an object and the StudentDatabaseApp.java file where the main function will be. In the Student.java file I import java.util.Scanner, then I create a class call Student with 7 private variables which is an example of encapsulation. Encapsulation is the process of wrapping code and data together into a single unit. All the data classes are private and a setmethod and getmethod are used to get access to the data, this makes class read-only and write-only. Encapsulate class is easy to test. Access modifiers, there are two types of modifiers in java, access modifiers and non-access modifiers. The access modifiers in java specifies accessiblity, scope, of a data member, method, constructor or class. There are 4 types: private, default, protected, and public. Static and abstract are  a non-access modifiers.