Class, objects, and constructor
The student class or the cookie cutter is not an object and is empty and holds no space or memory but when the student cookie cutter is placed on the dough the constructor is activated and the student object is created. An object is any entity that has state and behavior. A state is not static it is something that changes. Static is for classes not for objects. A class are templates that are used to create objects and to define object data types and methods. A constructor is a block of codes similar to the method. It is called when an instnce of the object is created, and memory is allocated for the object. It is a special type of method which is used to initialize the object. A constructor is created at compile time and a written constructor is not needed because Java will automatically make one for the developer if one is not written. The constructor automatically created is the one with no parameter and it is default