Static member is add to the student class is added in 'private static int id'
In Java, a static member is a member of a class that isn't associated with an instance of a class. Instead, the member belongs to the class itself. As a result, you can access the static member without first creating a class instance. ... The value of a static field is the same across all instances of the class. A constructor cannot be static, abstract, final and synchronized