The enroll method - Enroll in Courses
A 'do' loop is created - The Java do-while loop is used to iterate a part of the program several times. If the number of iteration is not fixed and you must have to execute the loop at least once, it is recommended to use do-while loop. The Java do-while loop is executed at least once because condition is checked after loop body.
Inside the loop Students are instructed to enter the course to enroll in then the Scanner System.in method gets the input and moves to the next line after student press enters due to the 'in.nextLine()' method