Coordinates class
Created a class called coordinates along with two private final int variables called x and y. Then I created a public constructor called the same name as the java class with a int x and int y parameter and within its bracket I am using the keyword this to address and use the recent or current x and y variables and below that there are two methods. One is named getX with a public int type and the other method is called getY with a public int databype.