Created folders and TestConnection.java file, input code
The will be the top section of the 3 stack model I created for myself so I could understand MVC. The top part will like the introduction and is the model which will have the JDBC, DAO, Utilities, models with its POJOs. The sql driver being used is the postgresql.driver. Java Dababase Connectivity API is used to connet to and exchange data with databases.