add driver jar to build path
This driver is a piece of software that lets our jave application connect to and exchange commands and results with our database. The org.postgressql drive class is defined in the postgresql JDBC JAR. The jar file needs to be included in the build path, you have to download the jar file.