Prepared statement
The prepared statement which is a pre-compiled statement that has placeholders for specific data value is used whenever the same SQL command is to be executed several times. Since it is precompiled, only the data values are changed during runtime. Prepared statements can help protect your application against a malicious attack called sql injection. To create a prepared statement the prepareStatement method from our Connection object will be used. The statement will be passed as a String