prepared statement connection
To create a prepared statement we will use the prepareStatement method from our Connection object. We will need to pass in our statement as a string. There is a Student table that a row will be inserted into. I am going to insert a row into the STUDENTS table but the values will be replaced with question marks