stmt1.setInt(1, 1)
I set the values in order. The first parameter is the index of the placeholder and the second parameter is the value passed in. To execute this statement we call the stmt1.executeUpdate() method To check that. within the while loop we can retrieve columns either by their index value or by their name